html/semantics/embedded-content/the-video-element/video-loading-lazy-poster-when-visible.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #target {
    position: absolute;
    top: 1000vh;
  }
</style>

<video id="target" poster="/media/2048x1360-random.jpg" loading="lazy"></video>

<script>
  const video = document.getElementById("target");

  const t = async_test(
    "Video with loading=lazy does not fetch poster while not visible, then fetches poster once scrolled into view"
  );

  // Step 1: Not visible => check initial dimensions.
  t.step_timeout(() => {
    t.step(() => {
      const initialWidth = video.offsetWidth;
      const initialHeight = video.offsetHeight;

      // Step 2: Scroll into view => poster should load and dimensions should change.
      video.scrollIntoView();

      t.step_timeout(() => {
        t.step(() => {
          const finalWidth = video.offsetWidth;
          const finalHeight = video.offsetHeight;
          assert_not_equals(
            finalWidth,
            initialWidth,
            "Video width should change after poster loads"
          );
          assert_not_equals(
            finalHeight,
            initialHeight,
            "Video height should change after poster loads"
          );
        });
        t.done();
      }, 2000);
    });
  }, 2000);
</script>
</body>
</html>
